Imported Debian patch 2:3.6.6-6+deb7u7

+++ lib/util/data_blob.c
@@ -18,9 +18,7 @@
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "replace.h"
-#include "attr.h"
-#include "data_blob.h"
+#include "includes.h"
 
 const DATA_BLOB data_blob_null = { NULL, 0 };
 
@@ -135,7 +133,7 @@ print the data_blob as hex string
 **/
 _PUBLIC_ char *data_blob_hex_string_lower(TALLOC_CTX *mem_ctx, const DATA_BLOB *blob)
 {
-	size_t i;
+	int i;
 	char *hex_string;
 
 	hex_string = talloc_array(mem_ctx, char, (blob->length*2)+1);
@@ -155,7 +153,7 @@ _PUBLIC_ char *data_blob_hex_string_lower(TALLOC_CTX *mem_ctx, const DATA_BLOB *
 
 _PUBLIC_ char *data_blob_hex_string_upper(TALLOC_CTX *mem_ctx, const DATA_BLOB *blob)
 {
-	size_t i;
+	int i;
 	char *hex_string;
 
 	hex_string = talloc_array(mem_ctx, char, (blob->length*2)+1);
